runt control is working with python library, pwm speed control is also supported
