+++ mobility/autonomous/runt_test.js
@@ -1,0 +1,3 @@
+var rover = require('./python_control.js');
+
+rover.turn_right();
