--- mobility/runt_test.js
@@ -1,14 +1,0 @@
-
-var options = {
-  mode: 'text',
-  pythonPath: '/usr/bin/python',
-  pythonOptions: ['-u'],
-  scriptPath: './',
-  args: ['f']
-};
-
-PythonShell.run('drive_control.py', options, function (err, results) {
-  if (err) throw err;
-  // results is an array consisting of messages collected during execution
-  console.log('results: %j', results);
-});
